Add Dynamo 2.1 Dropdown Unit Tests
